fix: 优化GUI
1. 使用threading，防止点击下载按钮后，同步执行的下载操作卡住主进程
2. 将图片文件的创建修改为`debug=True`时才进行，方便GUI情况不创建临时目录，所有信息全部由内存管理。并移除GUI中临时目录的创建，删除操作。
3. 抛出 NsfcDownloader 的更多细节，使得GUI能够通过对象属性进行状态更新
4. 增加 国家科学基金共享服务网 快捷页面链接，增加项目声明。
5. 补充README.md细节

+++ nsfc_downloader.py
@@ -12,12 +12,16 @@ __AUTHOR__ = 'Rhilip'
 
 
 def arg_parser():
-    parse = argparse.ArgumentParser(
-        description='A tool to Download PDF format conclusion from http://output.nsfc.gov.cn/')
-    parse.add_argument('--ratify', '-r', help='The ratify id of the project you want to download', required=True)
-    parse.add_argument('--tmp_path', '-t', default='./tmp', help='The path you want to save tmp file')
-    parse.add_argument('--out_path', '-o', default='./output', help='The path you want to save output PDF file')
-    return parse.parse_args()
+    parser = argparse.ArgumentParser(
+        description='A tool to Download PDF format conclusion from http://output.nsfc.gov.cn/'
+    )
+    parser.add_argument('--version', '-v', action='version', version=__VERSION__)
+    parser.add_argument('--ratify', '-r', help='The ratify id of the project you want to download', required=True)
+    parser.add_argument('--tmp_path', '-t', default='./tmp', help='The path you want to save tmp file')
+    parser.add_argument('--out_path', '-o', default='./output', help='The path you want to save output PDF file')
+    parser.add_argument('--no-debug', dest='debug', action='store_false', help='Disable The debug mode')
+    parser.set_defaults(debug=True)
+    return parser.parse_args()
 
 
 def mkdir_p(path, mode=0o777):
@@ -29,6 +33,7 @@ def mkdir_p(path, mode=0o777):
     :return:
     """
     try:
+        path = os.path.abspath(path)
         os.makedirs(path, mode)
     except OSError as exc:
         if exc.errno == errno.EEXIST and os.path.isdir(path):
@@ -57,16 +62,25 @@ def clean_filename(string: str) -> str:
 
 
 class NsfcDownloader:
+    i = 1
+    download_finish = False
+    download_stats = {}
+
     debug = True
     ratify_info = {}
 
-    def __init__(self, tmp_path, out_path):
-        self.tmp_path = tmp_path
+    def __init__(self, out_path, tmp_path):
         self.out_path = out_path
+        self.tmp_path = tmp_path
 
         # 使得相关文件夹存在
-        mkdir_p(tmp_path)
         mkdir_p(out_path)
+        mkdir_p(tmp_path)
+
+    def clear_state(self):
+        self.i = 1
+        self.download_finish = False
+        self.download_stats = {}
 
     def get_ratify_info_from_nsfc(self, ratify) -> dict:
         """
@@ -80,8 +94,9 @@ class NsfcDownloader:
         try:
             r.raise_for_status()
             rj = r.json()  # project_info
-        except requests.HTTPError as e:
-            rj = {'success': False, 'code': r.status_code}
+            rj['success'] = True
+        except Exception as e:
+            rj = {'success': False, 'msg': str(e)}
             # Debug 模式下重新抛出抛出
             if self.debug:
                 raise e
@@ -115,24 +130,28 @@ class NsfcDownloader:
 
         ratify_prefix = ratify[:2]
 
-        i = 1
+        self.i = 1
         while True:
-            tmp_file = os.path.join(self.tmp_path, '{}_{}.png'.format(ratify, i))
+            tmp_file = os.path.join(self.tmp_path, '{}_{}.png'.format(ratify, self.i))
             if os.path.exists(tmp_file):
                 content = open(tmp_file, 'rb').read()
             else:
-                req_url = "http://output.nsfc.gov.cn/report/{}/{}_{}.png".format(ratify_prefix, ratify, i)
+                req_url = "http://output.nsfc.gov.cn/report/{}/{}_{}.png".format(ratify_prefix, ratify, self.i)
                 print('正在请求页面 {}'.format(req_url))
                 r = requests.get(req_url, timeout=10)
                 if r.status_code == 404:
                     break
                 content = r.content
-                with open(tmp_file, 'wb') as tmp_f:
-                    tmp_f.write(r.content)
+
+                if self.debug:
+                    with open(tmp_file, 'wb') as tmp_f:
+                        tmp_f.write(r.content)
 
             img_files_list.append(tmp_file)
             img_bytes_list.append(content)
-            i += 1
+            self.i += 1
+
+        self.download_finish = True
 
         return img_files_list, img_bytes_list
 
@@ -148,32 +167,38 @@ class NsfcDownloader:
             status['path'] = os.path.join(self.out_path, clean_filename('{} {}.pdf'.format(ratify, project_name)))
 
             if os.path.exists(status['path']):
+                status['success'] = True
                 status['msg'] = 'PDF已存在，请打开 `{}`。'.format(status['path'])
             else:
-                print('开始下载 {} {}'.format(ratify, project_name))
-                img_files_list, img_bytes_list = self.download_loop(ratify)
-                print('下载完成 {} {}'.format(ratify, project_name))
+                try:
+                    print('开始下载 {} {}'.format(ratify, project_name))
+                    img_files_list, img_bytes_list = self.download_loop(ratify)
+                    print('下载完成 {} {}'.format(ratify, project_name))
 
-                if len(img_bytes_list) > 0:
-                    print('正在组合PDF {}'.format(status['path']))
-                    pdf = img2pdf.convert(img_bytes_list)
-                    with open(status['path'], "wb") as file_:
-                        file_.write(pdf)
-                        status['success'] = True
+                    if len(img_bytes_list) > 0:
+                        print('正在组合PDF {}'.format(status['path']))
+                        pdf = img2pdf.convert(img_bytes_list)
+                        with open(status['path'], "wb") as file_:
+                            file_.write(pdf)
+                            status['success'] = True
 
-                    print('移除临时文件')
-                    for f in img_files_list:
-                        os.remove(f)
-                else:
-                    status['msg'] = '下载过程出现问题，未获得有效图片。'
+                        if self.debug:
+                            print('移除临时文件')
+                            for f in img_files_list:
+                                os.remove(f)
+                    else:
+                        status['msg'] = '下载过程出现问题，未获得有效图片。'
+                except Exception as e:
+                    status['msg'] = '内部错误： {}'.format(e)
 
         if status['msg']:
             print(status['msg'])
 
-        return status
+        self.download_stats = status
+        return self.download_stats
 
 
 if __name__ == '__main__':
     args = arg_parser()
-    downloader = NsfcDownloader(args.tmp_path, args.out_path)
+    downloader = NsfcDownloader(args.out_path, args.tmp_path)
     downloader.download(args.ratify)
